New files

+++ Assignment/nodejs/User_input.js
@@ -4,7 +4,7 @@ const readline = require("readline").createInterface({
 });
 
 readline.question("Enter a name:", (input1) => {
-  const name1 = parseFloat(input1);
+  const name1 = parseInt(input1);
     console.log(input1);
     readline.close();
 });
